+++ L2J_Mobius_C6_Interlude/java/com/l2jmobius/gameserver/geodata/geodriver/blocks/MultilayerBlock.java
@@ -107,7 +107,7 @@ public class MultilayerBlock implements IBlock
 	
 	private int _extractLayerNswe(short layer)
 	{
-		return (byte) (layer & 0x000F);
+		return (layer & 0x000F);
 	}
 	
 	private int _extractLayerHeight(short layer)
