+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/instancemanager/FourSepulchersManager.java
@@ -1410,7 +1410,7 @@ public final class FourSepulchersManager
 	
 	protected byte minuteSelect(byte min)
 	{
-		if (((double) min % 5) != 0) // if doesn't divides on 5 fully
+		if ((min % 5) != 0) // if doesn't divides on 5 fully
 		{
 			// mad table for selecting proper minutes...
 			// may be there is a better way to do this
