+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/network/serverpackets/SSQStatus.java
@@ -137,8 +137,8 @@ public class SSQStatus implements IClientOutgoingPacket
 				
 				if (totalOverallScore != 0)
 				{
-					dawnPercent = Math.round(((float) dawnTotalScore / (float) totalOverallScore) * 100);
-					duskPercent = Math.round(((float) duskTotalScore / (float) totalOverallScore) * 100);
+					dawnPercent = Math.round(((float) dawnTotalScore / totalOverallScore) * 100);
+					duskPercent = Math.round(((float) duskTotalScore / totalOverallScore) * 100);
 				}
 				
 				/* DUSK */
@@ -238,20 +238,20 @@ public class SSQStatus implements IClientOutgoingPacket
 						else
 						{
 							packet.writeC(0);
-							packet.writeC(Math.round(((float) dawnProportion / (float) totalDawnMembers) * 100));
+							packet.writeC(Math.round(((float) dawnProportion / totalDawnMembers) * 100));
 						}
 					}
 					else
 					{
 						if (totalDawnMembers == 0)
 						{
-							packet.writeC(Math.round(((float) duskProportion / (float) totalDuskMembers) * 100));
+							packet.writeC(Math.round(((float) duskProportion / totalDuskMembers) * 100));
 							packet.writeC(0);
 						}
 						else
 						{
-							packet.writeC(Math.round(((float) duskProportion / (float) totalDuskMembers) * 100));
-							packet.writeC(Math.round(((float) dawnProportion / (float) totalDawnMembers) * 100));
+							packet.writeC(Math.round(((float) duskProportion / totalDuskMembers) * 100));
+							packet.writeC(Math.round(((float) dawnProportion / totalDawnMembers) * 100));
 						}
 					}
 				}
